+++ scripts/automation/tests/nose_helper.py
@@ -18,7 +18,7 @@ def get_nose_runner(report_folder, parallel=True, process_timeout=600, process_r
                 or not os.path.isdir(report_folder):
             raise ValueError('Report folder {} does not exist'.format(report_folder))
 
-        arguments = [__file__, '-v']
+        arguments = [__file__, '-v', '--nologcapture']
         if parallel:
             arguments += ['--processes=-1', '--process-timeout={}'.format(process_timeout)]
             if process_restart:
@@ -35,6 +35,7 @@ def get_nose_runner(report_folder, parallel=True, process_timeout=600, process_r
 
         debug_file = os.path.join(report_folder, 'test-debug.log')
         arguments += ['--debug-log={}'.format(debug_file)]
+        arguments += ['--nologcapture']
         arguments.extend(test_folders)
 
         result = nose.run(argv=arguments)