+++ scripts/jenkins_perf.sh
@@ -2,8 +2,12 @@
 
 # Run performance in jenkins build
 
+# Activate virtual environment
 python -m virtualenv --clear env
 . ./env/bin/activate
+
+# Load command build variables
+. $(cd $(dirname $0); pwd)/jenkins_common.sh
 
 echo "Run performance test on $(hostname)"
 
@@ -12,17 +16,11 @@ if [ -z %BUILD_NUMBER ]; then
     exit 1
 fi
 
-version=$(printf '%.8d' $BUILD_NUMBER)
-echo "Version number: $version"
-
 echo 'Before install'
 which az
 pip list
 
-build_folder=/var/build_share/$BRANCH_NAME/$version
-echo "Install build from $build_folder"
-
-python -m pip install azure-cli --find-links file://$build_folder/artifacts/build -v
+python -m pip install azure-cli --find-links file://$build_share/artifacts/build -v
 python -m pip freeze
 
 python ./scripts/performance/measure.py