+++ src/command_modules/azure-cli-shell/azclishell/argfinder.py
@@ -5,6 +5,7 @@
 
 import argparse
 import os
+import io
 import sys
 
 from argcomplete import CompletionFinder
@@ -27,7 +28,7 @@ class ArgsFinder(CompletionFinder):
 
         try:
             stderr = sys.stderr
-            sys.stderr = os.open(os.devnull, "w")
+            sys.stderr = io.open(os.devnull, "w")
 
             active_parsers[0].parse_known_args(comp_words, namespace=parsed_args)